socialme/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
bce-python-sdk==0.9.17
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
blinker==1.8.2
boto3==1.34.133
botocore==1.34.133
cachetools==5.3.3
celery==5.4.0
cement==3.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.8
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
defusedxml==0.8.0rc2
diff-match-patch==20230430
distro==1.9.0
dj-database-url==2.2.0
Django==4.2.4
django-allauth==0.63.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-import-export==4.1.1
django-otp==1.5.0
django-storages==1.14.3
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.1
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
haversine==2.8.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib==3.9.1
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
oauth2==1.9.0.post1
oauthlib==3.2.2
odfpy==1.4.1
openai==1.37.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.1
paddle-bfloat==0.1.7
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pathspec==0.12.1
pdf2docx==0.5.8
pillow==10.3.0
platformdirs==4.2.2
premailer==3.10.0
prompt_toolkit==3.0.47
protobuf==3.20.2
psutil==6.0.0
psycopg2==2.9.9
pyclipper==1.3.0.post5
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
pyshorteners==1.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.9.4
rarfile==4.2
redis==5.0.7
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.14.0
sentry-sdk==2.11.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
tablib==3.5.0
termcolor==2.4.0
tifffile==2024.7.24
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
